refactor: addStatement 람다식 표현
