+++ src/main/java/dao/UserDao.java
@@ -1,12 +1,12 @@
 package dao;
 
+import dao.strategy.AddStatement;
+import dao.strategy.DeleteAllStatement;
+import dao.strategy.StatementStrategy;
 import domain.User;
 
 import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
+import java.sql.*;
 
 public class UserDao {
     private DataSource dataSource;
@@ -15,7 +15,25 @@ public class UserDao {
         this.dataSource = dataSource;
     }
 
-    public void jdbcContextWithStatementStrategy(StatementStrategy stmt) throws SQLException {
+    public void add(User user) throws SQLException {
+        StatementStrategy st = new AddStatement(user);
+        jdbcContextWithStatementStrategy(
+                c -> {
+                    PreparedStatement ps = c.prepareStatement("insert into users(id, name, password) values (?,?,?)");
+                    ps.setString(1, user.getId());
+                    ps.setString(2, user.getId());
+                    ps.setString(3, user.getId());
+                    return ps;
+                }
+        );
+    }
+
+    public void deleteAll() throws SQLException {
+        StatementStrategy st = new DeleteAllStatement();
+        jdbcContextWithStatementStrategy(st);
+    }
+
+    private void jdbcContextWithStatementStrategy(StatementStrategy stmt) throws SQLException {
         Connection c = null;
         PreparedStatement ps = null;
 
@@ -41,21 +59,6 @@ public class UserDao {
         }
     }
 
-    public void add(User user) throws SQLException {
-        Connection c = dataSource.getConnection();
-
-        PreparedStatement ps = c.prepareStatement(
-                "insert into users(id, name, password) values(?, ?, ?)"
-        );
-        ps.setString(1, user.getId());
-        ps.setString(2, user.getName());
-        ps.setString(3, user.getPassword());
-
-        ps.executeUpdate();
-        ps.close();
-        c.close();
-    }
-
     public User get(String id) throws SQLException {
         Connection c = dataSource.getConnection();
 
@@ -76,11 +79,6 @@ public class UserDao {
         c.close();
 
         return user;
-    }
-
-    public void deleteAll() throws SQLException {
-        StatementStrategy st = new DeleteAllStatement();
-        jdbcContextWithStatementStrategy(st);
     }
 
     public int getCount() throws SQLException {
